chore: split infer.rs into more files
