+++ crates/hir/src/diagnostics.rs
@@ -13,7 +13,8 @@ use hir_ty::{
     builtins::BuiltinId,
     database::HirDatabase,
     diagnostics::InferenceDiagnosticKind,
-    infer::{LoweredKind, TypeExpectation, TypeLoweringError, TypeLoweringErrorKind},
+    infer::TypeExpectation,
+    lower::{LoweredKind, TypeContainer, TypeLoweringError, TypeLoweringErrorKind},
     ty::Type,
     validate::AddressSpaceError,
 };
@@ -350,7 +351,7 @@ pub(crate) fn any_diag_from_infer_diagnostic(
         InferenceDiagnosticKind::InvalidType {
             error: TypeLoweringError { container, kind },
         } => match container {
-            hir_ty::infer::TypeContainer::Expression(expression) => {
+            TypeContainer::Expression(expression) => {
                 let pointer = source_map.expression_to_source(*expression).ok()?.clone();
                 let source = InFile::new(file_id, pointer);
 
@@ -359,7 +360,7 @@ pub(crate) fn any_diag_from_infer_diagnostic(
                     error: kind.clone(),
                 }
             },
-            hir_ty::infer::TypeContainer::TypeSpecifier(type_specifier) => {
+            TypeContainer::TypeSpecifier(type_specifier) => {
                 let pointer = source_map
                     .type_specifier_to_source(*type_specifier)
                     .ok()?
