+++ crates/hir_ty/src/lower/eval.rs
@@ -7,8 +7,9 @@ use wgsl_types::{
 };
 
 use crate::{
-    infer::{
+    lower::{
         Lowered, TypeContainer, TypeLoweringContext, TypeLoweringError, TypeLoweringErrorKind,
+        generics::{TemplateParameter, TemplateParameters},
     },
     ty::{Type, TypeKind},
 };
@@ -186,87 +187,6 @@ impl TypeLoweringContext<'_> {
             .map(|argument| (self.evaluate_template_argument(*argument), *argument))
             .collect();
         let length = template_parameters.len();
-        TemplateParameters {
-            container,
-            inner: template_parameters,
-            length,
-        }
-    }
-}
-
-/// A single template parameter.
-#[derive(Clone, Debug, PartialEq)]
-pub enum TemplateParameter {
-    Type(Type),
-    /// The error instance is encoded as a `None`.
-    Instance(Option<Instance>),
-    Enumerant(Enumerant),
-}
-
-pub struct TemplateParameters {
-    container: TypeContainer,
-    inner: VecDeque<(TemplateParameter, ExpressionId)>,
-    length: usize,
-}
-
-impl TemplateParameters {
-    pub fn has_next(&self) -> bool {
-        !self.inner.is_empty()
-    }
-
-    pub fn next(&mut self) -> Option<(TemplateParameter, ExpressionId)> {
-        self.inner.pop_front()
-    }
-
-    pub fn next_as_type(&mut self) -> Result<(Type, ExpressionId), TypeLoweringError> {
-        match self.next() {
-            Some((TemplateParameter::Type(r#type), id)) => Ok((r#type, id)),
-            Some((_, id)) => Err(TypeLoweringError {
-                container: TypeContainer::Expression(id),
-                kind: TypeLoweringErrorKind::UnexpectedTemplateArgument("a type".to_owned()),
-            }),
-            None => Err(TypeLoweringError {
-                container: self.container,
-                kind: TypeLoweringErrorKind::MissingTemplateArgument("a type".to_owned()),
-            }),
-        }
-    }
-
-    pub fn next_as_instance(
-        &mut self
-    ) -> Result<(Option<Instance>, ExpressionId), TypeLoweringError> {
-        match self.next() {
-            Some((TemplateParameter::Instance(instance), id)) => Ok((instance, id)),
-            Some((_, id)) => Err(TypeLoweringError {
-                container: TypeContainer::Expression(id),
-                kind: TypeLoweringErrorKind::UnexpectedTemplateArgument("an instance".to_owned()),
-            }),
-            None => Err(TypeLoweringError {
-                container: self.container,
-                kind: TypeLoweringErrorKind::MissingTemplateArgument("an instance".to_owned()),
-            }),
-        }
-    }
-
-    pub fn next_as_enumerant(&mut self) -> Result<(Enumerant, ExpressionId), TypeLoweringError> {
-        match self.next() {
-            Some((TemplateParameter::Enumerant(enumerant), id)) => Ok((enumerant, id)),
-            Some((_, id)) => Err(TypeLoweringError {
-                container: TypeContainer::Expression(id),
-                kind: TypeLoweringErrorKind::UnexpectedTemplateArgument("an enum".to_owned()),
-            }),
-            None => Err(TypeLoweringError {
-                container: self.container,
-                kind: TypeLoweringErrorKind::MissingTemplateArgument("an enum".to_owned()),
-            }),
-        }
-    }
-
-    pub(crate) const fn len(&self) -> usize {
-        self.length
-    }
-
-    pub const fn container(&self) -> &TypeContainer {
-        &self.container
+        TemplateParameters::new(container, template_parameters)
     }
 }
